on click listener has been added to recycler view
