+++ app/src/main/res/layout/fragment_book.xml
@@ -9,8 +9,15 @@
     <!-- TODO: Update blank fragment layout -->
     <TextView
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:text="Book page"
+        android:layout_height="wrap_content"
+        android:text="Books"
+        android:gravity="center"
+        android:textSize="40sp"
         android:textColor="@color/black"/>
+    <androidx.recyclerview.widget.RecyclerView
+        android:id="@+id/recyclerview"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="55dp"/>
 
 </FrameLayout>
