+++ app/src/main/res/layout/fragment_home.xml
@@ -23,10 +23,7 @@
         android:layout_marginTop="20dp"
         android:textSize="25sp"
         android:textColor="@color/black"/>
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/recyclerview"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"/>
+
