+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -11,9 +11,6 @@
         android:icon="@drawable/ic_baseline_menu_book_24"
         android:title="Books" />
 
-    <item
-        android:id="@+id/profile"
-        android:icon="@drawable/ic_profile1"
-        android:title="Profile" />
+ 
 
 </menu>
